Fix List type accepting strings
Currently if you pass a string value in place of list,
it will validate OK, because strings in Python are also
sequences (of individual char strings). From validation
point of view it is incorrect.

This patch adds extra check to List type do check that
values are not strings (but you still can use custom
sequence types).