Fix Dict missing keys validation on load
If Dict is defined with fixed keys, there is a
mechanism that first validates whatever keys are
present and then tries to load all other fixed
keys (hoping for example for a default value).
The bug in that code triggered double validtion
for existing keys which values had errors.

Fixed now.